Modified Player Verification
Before this, it was possible that the JWT token was valid for a user's requests and therefore that user could do all actions within the system.

 However, this raises a problem, if the user is deleted or changed, whilst in the database that change has occured, the JWT token is still valid and therefore that user could still perform those actions.

So now when the token is validated we also check if the player exists as is claimed in the database. This essentially nullifies the need for the Verify endpoint which was also faulty, however, it is being kept because the game uses it and therefore no changes need to be made on the frontend.