Full Refactor (Cache)
Added cache for all service related aspects.

Refactored all services to implement this.

Major upsides: reduced amount of queries to db by very large margins, example for 200 items the service was doing 801 queries

1 (load items)
+ N (author)
+ N (rating)
+ N (subscription check)
+ N (level/machine)
= 1 + 4N queries

which means that if we had 10 items and refreshed a page 20 times, we would do 1020 queries to the DB, impacting compute usage greatly, now we are doing 0 queries on the best case scenario and 4 worst case.