Password Hashing
Passwords are now non reversible to ensure more security. Before, they could be decrypted which meant that anyone with the Key, Salt and the encrypted password could reverse them.

While this could only be done with acess to environment variables and the database,  it is safer to store passwords as non reversible, so this change should ensure that.

For now, old passwords are migrated and therefore the AesCryptoService is still in use.